+++ src/aioway/__init__.py
@@ -1,0 +1,1 @@
+# Copyright (c) AIoWay Authors - All Rights Reserved
