+++ src/aioway/app/entrypoint.py
@@ -6,9 +6,8 @@ from collections import abc as cabc
 from torch import nn, optim
 
 from aioway.dsets import Dset, route_dset
-from aioway.emits import emit, route_loss
+from aioway.nn import TSpec, as_tspec, emit, route_loss
 from aioway.trainers import StaticTrainer, TrainCfg
-from aioway.nn import TSpec, as_tspec
 
 from .server import serve
