+++ src/aioway/nn/__init__.py
@@ -1,4 +1,5 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
 from .deductions import *
+from .emits import *
 from .tspecs import *
