+++ src/aioway/nn/emits/_utils.py
