+++ src/aioway/nn/emits/compound.py
