+++ src/aioway/nn/emits/linear.py
