+++ src/aioway/nn/emits/losses/contrastive.py
@@ -8,8 +8,8 @@ import torch
 from torch import nn, optim
 from torch.nn import functional as F
 
-from aioway.emits import Emitter, emitter_dcls
 from aioway.nn import TSpec
+from aioway.nn.emits import Emitter, emitter_dcls
 
 __all__ = ["ContrastiveLoss", "ContrastiveLossEmitter"]
