+++ src/aioway/nn/emits/losses/routes.py
@@ -7,7 +7,6 @@ from collections import abc as cabc
 from torch import nn
 from torchrl.data import tensor_specs as tspecs
 
-from aioway.emits import emitter_function
 from aioway.nn import (
     ArgsTSpec,
     LossTSpec,
@@ -16,6 +15,7 @@ from aioway.nn import (
     as_tspec,
     is_tspec_like,
 )
+from aioway.nn.emits.emitters import emitter_function
 
 __all__ = ["emit_loss", "route_loss"]
