+++ tests/nn/emits/losses/test_losses.py
@@ -4,8 +4,7 @@ import pytest
 from torch import nn
 from torchrl.data import tensor_specs as tspecs
 
-from aioway.emits import emit_loss, emit_one, route_loss
-from aioway.nn import ArgsTSpec, LossTSpec
+from aioway.nn import ArgsTSpec, LossTSpec, emit_loss, emit_one, route_loss
 
 
 @pytest.fixture
