+++ tests/nn/emits/test_compound.py
@@ -4,7 +4,7 @@ import pytest
 import torch
 from torch import nn
 
-from aioway.emits import BuiltModule, CompoundBuilder
+from aioway.nn import BuiltModule, CompoundBuilder
 
 
 @pytest.fixture
