+++ tests/nn/emits/test_convs.py
@@ -3,16 +3,16 @@
 import pytest
 from torch import nn
 
-from aioway.emits import emit_one
-from aioway.emits.convs import ImageRegressorEmitter
-from aioway.schemas import Shape
 from aioway.nn import (
     TSpec,
+    emit_one,
     float_image_tspec,
     sample_from_tspec,
     set_batch_size,
     unbounded_box_tspec,
 )
+from aioway.nn.emits.convs import ImageRegressorEmitter
+from aioway.schemas import Shape
 
 
 @pytest.fixture
