+++ tests/nn/emits/test_linear.py
@@ -8,15 +8,16 @@ from torch import nn
 from torch.utils import data as dutils
 from torchrl.data import tensor_specs as tspecs
 
-from aioway.emits import (
+from aioway.nn import (
     MlpCompoundEmitter,
     MlpEmitter,
+    TSpec,
     emit,
     emit_one,
     linear_regression,
+    unbounded_box_tspec,
 )
 from aioway.schemas import Shape
-from aioway.nn import TSpec, unbounded_box_tspec
 
 
 @pytest.fixture
