+++ tests/nn/emits/test_norms.py
@@ -5,8 +5,13 @@ import typing
 import pytest
 import torch
 
-from aioway.emits import NormEmitter, NormType, emit_one, layer_norm_emitter
-from aioway.nn import unbounded_box_tspec
+from aioway.nn import (
+    NormEmitter,
+    NormType,
+    emit_one,
+    layer_norm_emitter,
+    unbounded_box_tspec,
+)
 
 
 @pytest.fixture(params=typing.get_args(NormType.__value__))
